# swagger.yaml
openapi: 3.0.1
info:
  title: componentsCRUD
  description: primera version
  version: "1.0"
servers:
- url: http://localhost:8084
  description: Generated server url
paths:
  /api/patient/{id}:
    get:
      tags:
      - Patient
      summary: list Patient ForId
      operationId: listPatientForId
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PatientDTO'
    put:
      tags:
      - Patient
      summary: Update Patient
      operationId: updatePatient
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatientDTO'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PatientDTO'
    delete:
      tags:
      - Patient
      summary: Delete Patient
      operationId: deletePatient
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
  /api/doctor/{id}:
    get:
      tags:
      - Doctor
      summary: List Doctor For Id
      operationId: listDoctorForId
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/DoctorDTO'
    put:
      tags:
      - Doctor
      summary: Update Doctor
      operationId: updateDoctor
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DoctorDTO'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/DoctorDTO'
    delete:
      tags:
      - Doctor
      summary: Delete Doctor
      operationId: deleteDoctor
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
  /api/consult/{id}:
    get:
      tags:
      - Consult
      summary: Consult for id
      operationId: getConsult
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ConsultDTO'
    put:
      tags:
      - Consult
      summary: Update Consult
      operationId: updateConsult
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConsultDTO'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ConsultDTO'
    delete:
      tags:
      - Consult
      summary: Delete Consult
      operationId: deleteConsult
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
  /api/cite/{id}:
    get:
      tags:
      - Cite
      summary: get all cite for id
      operationId: listCitePorId
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CiteDTO'
    put:
      tags:
      - Cite
      summary: update cite
      operationId: updateCite
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CiteDTO'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CiteDTO'
    delete:
      tags:
      - Cite
      summary: delete cite
      operationId: deleteCite
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
  /api/patient:
    get:
      tags:
      - Patient
      summary: List Patient
      operationId: listPatient
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PatientDTO'
    post:
      tags:
      - Patient
      summary: Save Patient
      operationId: savePatient
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatientDTO'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PatientDTO'
  /api/doctor:
    get:
      tags:
      - Doctor
      summary: Consult for id
      operationId: listDoctor
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DoctorDTO'
    post:
      tags:
      - Doctor
      summary: Save Doctor
      operationId: saveDoctor
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DoctorDTO'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/DoctorDTO'
  /api/consult:
    get:
      tags:
      - Consult
      summary: list Consult
      operationId: listConsult
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ConsultDTO'
    post:
      tags:
      - Consult
      summary: Save Consult
      operationId: saveConsult
      parameters:
      - name: citeId
        in: query
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConsultDTO'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ConsultDTO'
  /api/cite/{idPatient}/{idDoctor}:
    post:
      tags:
      - Cite
      summary: save cite
      operationId: saveCite
      parameters:
      - name: idPatient
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: idDoctor
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CiteDTO'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CiteDTO'
  /api/patient/{id}/cite:
    get:
      tags:
      - Patient
      summary: List Cite For PatientId
      operationId: listCiteForPatientId
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CiteDTO'
  /api/doctor/{id}/cite:
    get:
      tags:
      - Doctor
      summary: list Cite For Doctor Id
      operationId: listCiteForDoctorId
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CiteDTO'
  /api/doctor/specialty/{specialty}:
    get:
      tags:
      - Doctor
      summary: Lis Doctor For Specialty
      operationId: lisDoctorForSpecialty
      parameters:
      - name: specialty
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DoctorDTO'
  /api/consult/search:
    get:
      tags:
      - Consult
      summary: list Consult for Info
      operationId: listConsultPorInfo
      parameters:
      - name: searchTerm
        in: query
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ConsultDTO'
  /api/consult/cite/{citeId}:
    get:
      tags:
      - Consult
      summary: list Consult For Cite
      operationId: listConsultForCite
      parameters:
      - name: citeId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ConsultDTO'
  /api/cite:
    get:
      tags:
      - Cite
      summary: get all cite
      operationId: listCite
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CiteDTO'
  /api/cite/status/{statusCite}:
    get:
      tags:
      - Cite
      summary: list Cite For Status
      operationId: listCiteForStatus
      parameters:
      - name: statusCite
        in: path
        required: true
        schema:
          type: string
          enum:
          - PENDIENTE
          - CANCELADA
          - REALIZADA
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CiteDTO'
  /api/cite/patient/{PatientId}:
    get:
      tags:
      - Cite
      summary: list Cite For PatientId
      operationId: listCiteForPatientId_1
      parameters:
      - name: PatientId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CiteDTO'
  /api/cite/doctor/{doctorId}:
    get:
      tags:
      - Cite
      summary: list Cite For DoctorId
      operationId: listCiteForDoctorId_1
      parameters:
      - name: doctorId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CiteDTO'
components:
  schemas:
    PatientDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        birthDate:
          type: string
          format: date-time
        disease:
          type: boolean
    DoctorDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        email:
          type: string
        specialty:
          type: string
    CiteDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        dateCite:
          type: string
        statusCite:
          type: string
        cancel:
          type: boolean
        patientId:
          type: integer
          format: int64
        doctorId:
          type: integer
          format: int64
    ConsultDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        dateConsult:
          type: string
        report:
          type: string
        citeDTO:
          $ref: '#/components/schemas/CiteDTO'
        dateConsultAsDate:
          type: string
          format: date-time
        dateConsultFromDate:
          type: string
          format: date-time
          writeOnly: true